requirements.txt
Flask==2.0.1
PyMySQL==1.0.2
Flask-Login==0.5.0
python-dotenv==0.19.1
Werkzeug==2.0.1
bcrypt==3.2.0
Markdown==3.3.4
python-slugify==5.0.2
Flask-WTF==1.0.0
WTForms==3.0.0
email-validator==1.1.3 